app/requirements.txt
Pillow>=10.0.0
segno>=1.6.1
